Redirect classic editor users to wp-admin after they confirm they really want to use classic editor
