+++ client/post-editor/editor-gutenberg-blocks-warning-dialog/index.jsx
@@ -4,7 +4,7 @@
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { identity, noop } from 'lodash';
+import { identity, noop, pickBy } from 'lodash';
 import { localize } from 'i18n-calypso';
 
 /**
@@ -15,6 +15,7 @@ import { Dialog } from '@automattic/components';
 import { setSelectedEditor } from 'state/selected-editor/actions';
 import { getSelectedSiteId } from 'state/ui/selectors';
 import { getEditedPostValue } from 'state/posts/selectors';
+import { getSiteAdminUrl } from 'state/sites/selectors';
 import getGutenbergEditorUrl from 'state/selectors/get-gutenberg-editor-url';
 import { openPostRevisionsDialog } from 'state/posts/revisions/actions';
 import {
@@ -25,6 +26,10 @@ import {
 	bumpStat,
 } from 'state/analytics/actions';
 import isGutenbergOptInEnabled from 'state/selectors/is-gutenberg-opt-in-enabled';
+import isSiteAutomatedTransfer from 'state/selectors/is-site-automated-transfer';
+import isPrivateSite from 'state/selectors/is-private-site';
+import { addQueryArgs } from 'lib/route';
+import wpcom from 'lib/wp';
 
 /**
  * Style dependencies
@@ -33,14 +38,19 @@ import './style.scss';
 
 class EditorGutenbergBlocksWarningDialog extends Component {
 	static propTypes = {
+		isAtomic: PropTypes.bool,
+		isPrivate: PropTypes.bool,
 		translate: PropTypes.func,
 		postContent: PropTypes.string,
 		siteId: PropTypes.number,
+		postId: PropTypes.number,
+		postType: PropTypes.string,
 		gutenbergUrl: PropTypes.string,
 		switchToGutenberg: PropTypes.func,
 		openPostRevisionsDialog: PropTypes.func,
 		optInEnabled: PropTypes.bool,
 		useClassic: PropTypes.func,
+		buildSiteAdminUrl: PropTypes.func,
 	};
 
 	static defaultProps = {
@@ -58,6 +68,12 @@ class EditorGutenbergBlocksWarningDialog extends Component {
 		isDialogVisible: false,
 		forceClassic: false,
 	};
+
+	componentDidMount() {
+		if ( ! this.props.optInEnabled ) {
+			this.redirectToWpAdmin();
+		}
+	}
 
 	shouldComponentUpdate( nextProps, nextState ) {
 		return this.state.isDialogVisible !== nextState.isDialogVisible;
@@ -78,9 +94,29 @@ class EditorGutenbergBlocksWarningDialog extends Component {
 
 	useClassicEditor = () => {
 		this.props.useClassic();
-		this.setState( {
-			forceClassic: true,
-		} );
+		this.redirectToWpAdmin();
+	};
+
+	redirectToWpAdmin = () => {
+		const { isAtomic, isPrivate, postId, postType, buildSiteAdminUrl } = this.props;
+		if ( isAtomic && isPrivate ) {
+			let queryArgs = pickBy( {
+				post: postId,
+				action: postId && 'edit', // If postId is set, open edit view.
+				post_type: postType !== 'post' && postType, // Use postType if it's different than post.
+				'classic-editor': 1,
+			} );
+
+			// needed for loading the editor in SU sessions
+			if ( wpcom.addSupportParams ) {
+				queryArgs = wpcom.addSupportParams( queryArgs );
+			}
+
+			const siteAdminUrl = buildSiteAdminUrl( postId ? 'post.php' : 'post-new.php' );
+			const wpAdminUrl = addQueryArgs( queryArgs, siteAdminUrl );
+
+			window.location.href = wpAdminUrl;
+		}
 	};
 
 	switchToGutenberg = () => {
@@ -186,11 +222,19 @@ export default connect( state => {
 	const postType = getEditedPostValue( state, siteId, postId, 'type' );
 	const gutenbergUrl = getGutenbergEditorUrl( state, siteId, postId, postType );
 	const optInEnabled = isGutenbergOptInEnabled( state, siteId );
+	const isAtomic = isSiteAutomatedTransfer( state, siteId );
+	const isPrivate = isPrivateSite( state, siteId );
 
 	return {
+		isAtomic,
+		isPrivate,
 		postContent,
 		siteId,
+		postId,
+		postType,
 		gutenbergUrl,
 		optInEnabled,
+		// eslint-disable-next-line wpcalypso/redux-no-bound-selectors
+		buildSiteAdminUrl: path => getSiteAdminUrl( state, siteId, path ),
 	};
 }, mapDispatchToProps )( localize( EditorGutenbergBlocksWarningDialog ) );
